IndigoAssets can now list the asset files
